% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_time_series.R
\name{build_time_series}
\alias{build_time_series}
\title{build_time_series}
\usage{
build_time_series(series)
}
\arguments{
\item{series}{}
}
\value{
A time series saved in object time_series for further use
}
\description{
build_time_series
}
\examples{

First, this command can download time-series of any frequency from Easydata of
the SBP. You just need to provide the object you saved using the download_series
function. Note: It is important to make sure that the series you have downloaded
does not have missing rows.

Considering Weighted-average overnight repo rate downloaded as WAORR.

This function can be used as:

build_time_series(WAORR)

The output will be a series called "time-series" to be used in subsequent analysis.

}
