% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download Dataset}
\usage{
download_dataset(series_ids, start_date, end_date, file_type = "csv")
}
\arguments{
\item{series_ids}{\itemize{
\item A vector containing series_keys for variables of interest
}}

\item{start_date}{\itemize{
\item Enter the start date for the intended series. If both start_date and end_date are not specified, the API will return only the most recent observation.
}}

\item{end_date}{\itemize{
\item Enter the end date for the intended series. If both start_date and end_date are not specified, the API will return only the most recent observation.
}}

\item{file_type}{\itemize{
\item Default is json, but csv is also available, whereby EasyData only provides a csv file
}}
}
\description{
Download Dataset
}
