% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_series.R
\name{download_series}
\alias{download_series}
\title{Download data csv file from EasyData}
\usage{
download_series(
  Series_ID,
  Easydata_Key,
  Start_date = NULL,
  End_date = NULL,
  File_type = NULL
)
}
\arguments{
\item{Series_ID}{\itemize{
\item Enter the Series ID from the EasyData platform
}}

\item{Easydata_Key}{\itemize{
\item Enter the "Easydata_key" already provided for the session
}}

\item{Start_date}{\itemize{
\item Enter the start date for the intended series. If both start_date and end_date are not specified, the API will return only the most recent observation.
}}

\item{End_date}{\itemize{
\item Enter the end date for the intended series. If both start_date and end_date are not specified, the API will return only the most recent observation.
}}

\item{File_type}{\itemize{
\item Default is json, but csv is also available, whereby EasyData only provides a csv file
}}
}
\value{
A csv file containing information on Series, Time Periods and Values of a variable of interest.
}
\description{
Download data csv file from EasyData
}
\examples{

Setup the EasyData Key

Easydata_key<-"C20D3D24160CE5693F56AA9846ABB2C438D8B112"

Read in Weighted-average Overnight Repo Rate

download_series("TS_GP_IR_REPOMR_D.ORR",Easydata_key, "2015-05-25" ,"2023-12-20", "csv")

}
