% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyData_key_setup.R
\name{EasyData_key_setup}
\alias{EasyData_key_setup}
\title{Entering in your EasyData API key for the current session}
\usage{
EasyData_key_setup(API_KEY)
}
\arguments{
\item{API_KEY}{\itemize{
\item Provide your EasyData API key here
}}
}
\value{
API key entered
}
\description{
Entering in your EasyData API key for the current session
}
\examples{

API_KEY <- "C11D3D29161CE5693F55AA9846ABB2C438D8B112"


EasyData_key_setup(API_KEY)

print(Easydata_key)

}
