% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Easydata_key.R
\name{get_Easydata_key}
\alias{get_Easydata_key}
\title{Retrieve Easydata API Key if already entered}
\usage{
get_Easydata_key()
}
\value{
The Easydata_key for the session if setup already; otherwise returns error message
}
\description{
Retrieve Easydata API Key if already entered
}
\examples{

get_Easydata_key()

}
