% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_on_series.R
\name{get_info_on_series}
\alias{get_info_on_series}
\title{Obtain custom information on a series}
\usage{
get_info_on_series(Series_ID, Information)
}
\arguments{
\item{Series_ID}{\itemize{
\item Enter the ID of the variable of interest
Input: Provide a string
}}

\item{Information}{\itemize{
\item Enter number 1 to 10 to retrieve information on the following:
}
\enumerate{
\item Reference Dataset Name
\item Series Short Name
\item Series Name
\item Series Description
\item Data Frequency
\item Unit
\item Variable Type
\item Available Since
\item Available Upto
\item Last Refresh Date
}

Input: All the above have to be entered as integers

Note: You need to have Easydata_key set before you can use this function}
}
\description{
Obtain custom information on a series
}
\examples{

They Easydata_key must be provided already by:

Easydata_key<-"C20D3D24160CE5693F56AA9846ABB2C438D8B112"

To obtain information on Available Since for Weighted-average Overnight Repo Rate, Enter:

get_info_on_series("TS_GP_IR_REPOMR_D.ORR",8)

}
