% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{plot_time_series}
\usage{
plot_time_series(time_series)
}
\arguments{
\item{time_series}{}
}
\value{
a graph of the time-series using ggplot2 library
}
\description{
plot_time_series
}
\examples{

You need to pass time_series built through the function build_time_series to
this function and it will plot the time-series.

plot_time_series(time_series)

}
