% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session_has_Easydata_key.R
\name{session_has_Easydata_key}
\alias{session_has_Easydata_key}
\title{Check to see if the current R session has an Easydata key setup}
\usage{
session_has_Easydata_key()
}
\value{
TRUE if the key is already read in as Easydata_key for the current session. Otherwise returns error message.
}
\description{
Check to see if the current R session has an Easydata key setup
}
\examples{

session_has_Easydata_key()

}
